Cloud functions code/Training/requirements.txt
azure-identity
azure-storage-blob
azure-cognitiveservices-vision-computervision
msrest
requests